Moved contents of compute_target.hpp into compute_target.cpp.
